runtime-rs: align Cloud Hypervisor CPU schema with v51.1
Cloud Hypervisor v51.1 models CPU counts as u32 and topology fields as
u16, but runtime-rs still used u8. With default_maxvcpus set to
MAX_CH_VCPUS (256), conversion failed with MaxVCPUsTooBig before the
config reached Cloud Hypervisor.

Widen CpusConfig, CpuTopology, and CpuAffinity to match the pinned CH
schema, drop the u8 try_from guards that encoded that mismatch, and add
a unit test for max_vcpus=256.

Fixes: #13449

+++ src/runtime-rs/crates/hypervisor/ch-config/src/lib.rs
@@ -62,14 +62,14 @@ pub enum ConsoleOutputMode {
 
 #[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
 pub struct CpuAffinity {
-    pub vcpu: u8,
-    pub host_cpus: Vec<u8>,
+    pub vcpu: u32,
+    pub host_cpus: Vec<usize>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
 pub struct CpusConfig {
-    pub boot_vcpus: u8,
-    pub max_vcpus: u8,
+    pub boot_vcpus: u32,
+    pub max_vcpus: u32,
     #[serde(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub topology: Option<CpuTopology>,
@@ -96,10 +96,10 @@ pub struct CpuFeatures {
 
 #[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
 pub struct CpuTopology {
-    pub threads_per_core: u8,
-    pub cores_per_die: u8,
-    pub dies_per_package: u8,
-    pub packages: u8,
+    pub threads_per_core: u16,
+    pub cores_per_die: u16,
+    pub dies_per_package: u16,
+    pub packages: u16,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, Default)]
